==> py/02_make-by-participant_data.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import glob

cleaned_data = pd.read_csv('data/02_cleaned/cleaned_data.csv')
cleaned_data_times = pd.read_csv('data/02_cleaned/cleaned_data_times.csv')

# make by-participant summaries

te_by_participants = cleaned_data.groupby(['participant', 'tense', 'emotion']).agg(
    mean_up = ('clicked_up', 'mean'),
    sum_up = ('clicked_up', 'sum'),
    mean_right = ('clicked_right', 'mean'),
    sum_right = ('clicked_right', 'sum'),
    depression_sum = ('depression_sum', 'mean'),
    anxiety_sum = ('anxiety_sum', 'mean'),
    stress_sum = ('stress_sum', 'mean'),
    n_item = ('item', 'nunique')
).reset_index()

t_by_participants = cleaned_data.groupby(['participant', 'tense']).agg(
    mean_up = ('clicked_up', 'mean'),
    sum_up = ('clicked_up', 'sum'),
    mean_right = ('clicked_right', 'mean'),
    sum_right = ('clicked_right', 'sum'),
    depression_sum = ('depression_sum', 'mean'),
    anxiety_sum = ('anxiety_sum', 'mean'),
    stress_sum = ('stress_sum', 'mean'),
    n_item = ('item', 'nunique')
).reset_index()

e_by_participants = cleaned_data.groupby(['participant', 'emotion']).agg(
    mean_up = ('clicked_up', 'mean'),
    sum_up = ('clicked_up', 'sum'),
    mean_right = ('clicked_right', 'mean'),
    sum_right = ('clicked_right', 'sum'),
    depression_sum = ('depression_sum', 'mean'),
    anxiety_sum = ('anxiety_sum', 'mean'),
    stress_sum = ('stress_sum', 'mean'),
    n_item = ('item', 'nunique')
).reset_index()

# reaction times ----

# make data
te_by_participants_rt = cleaned_data_times.groupby(['participant', 'tense', 'emotion']).agg(
    mean_rt = ('mouse_time_click', 'mean'),
    count = ('mouse_time_click', 'count'),
    n_item = ('item', 'nunique'),
    mean_depression = ('depression_sum', 'mean'),
    mean_anxiety = ('anxiety_sum', 'mean'),
    mean_stress = ('stress_sum', 'mean')
).reset_index()

t_by_participants_rt = cleaned_data_times.groupby(['participant', 'tense']).agg(
    mean_rt = ('mouse_time_click', 'mean'),
    count = ('mouse_time_click', 'count'),
    n_item = ('item', 'nunique'),
    mean_depression = ('depression_sum', 'mean'),
    mean_anxiety = ('anxiety_sum', 'mean'),
    mean_stress = ('stress_sum', 'mean')
).reset_index()

e_by_participants_rt = cleaned_data_times.groupby(['participant', 'emotion']).agg(
    mean_rt = ('mouse_time_click', 'mean'),
    count = ('mouse_time_click', 'count'),
    n_item = ('item', 'nunique'),
    mean_depression = ('depression_sum', 'mean'),
    mean_anxiety = ('anxiety_sum', 'mean'),
    mean_stress = ('stress_sum', 'mean')
).reset_index()

# by-participants data sets ----

# selections
te_by_participants.to_csv('data/02_cleaned/by_participants/te_by_participants.csv', encoding='utf-8', index=False)
t_by_participants.to_csv('data/02_cleaned/by_participants/t_by_participants.csv', encoding='utf-8', index=False)
e_by_participants.to_csv('data/02_cleaned/by_participants/e_by_participants.csv', encoding='utf-8', index=False)

# rts
te_by_participants_rt.to_csv('data/02_cleaned/by_participants/te_by_participants_rt.csv', encoding='utf-8', index=False)
t_by_participants_rt.to_csv('data/02_cleaned/by_participants/t_by_participants_rt.csv', encoding='utf-8', index=False)
e_by_participants_rt.to_csv('data/02_cleaned/by_participants/e_by_participants_rt.csv', encoding='utf-8', index=False)
