==> py/03_make-descriptive-statistics.ipynb <==
# Import libraries
import pandas as pd
import numpy as np

te_by_participants = pd.read_csv('data/02_cleaned/by_participants/te_by_participants.csv')
t_by_participants = pd.read_csv('data/02_cleaned/by_participants/t_by_participants.csv')
e_by_participants = pd.read_csv('data/02_cleaned/by_participants/e_by_participants.csv')

te_by_participants_rt = pd.read_csv('data/02_cleaned/by_participants/te_by_participants_rt.csv')
t_by_participants_rt = pd.read_csv('data/02_cleaned/by_participants/t_by_participants_rt.csv')
e_by_participants_rt = pd.read_csv('data/02_cleaned/by_participants/e_by_participants_rt.csv')

# make descriptives

te_descriptives = te_by_participants.groupby(['tense', 'emotion']).agg(
    mean_up = ('mean_up', 'mean'),
    sum_up = ('sum_up', 'sum'),
    mean_right = ('mean_right', 'mean'),
    sum_right = ('sum_right', 'sum'),
    n_participants = ('participant', 'nunique'),
    sum_item = ('n_item', 'sum')
).reset_index()

t_descriptives = t_by_participants.groupby(['tense']).agg(
    mean_up = ('mean_up', 'mean'),
    sum_up = ('sum_up', 'sum'),
    mean_right = ('mean_right', 'mean'),
    sum_right = ('sum_right', 'sum'),
    n_participants = ('participant', 'nunique'),
    sum_item = ('n_item', 'sum')
).reset_index()

e_descriptives = e_by_participants.groupby(['emotion']).agg(
    mean_up = ('mean_up', 'mean'),
    sum_up = ('sum_up', 'sum'),
    mean_right = ('mean_right', 'mean'),
    sum_right = ('sum_right', 'sum'),
    n_participants = ('participant', 'nunique'),
    sum_item = ('n_item', 'sum')
).reset_index()

# make descriptives rt

te_descriptives_rt = te_by_participants_rt.groupby(['tense', 'emotion']).agg(
    mean_rt = ('mean_rt', 'mean'),
    n_participants = ('participant', 'nunique'),
    sum_item = ('n_item', 'sum')
).reset_index()

t_descriptives_rt = t_by_participants_rt.groupby(['tense']).agg(
    mean_rt = ('mean_rt', 'mean'),
    n_participants = ('participant', 'nunique'),
    sum_item = ('n_item', 'sum')
).reset_index()

e_descriptives_rt = e_by_participants_rt.groupby(['emotion']).agg(
    mean_rt = ('mean_rt', 'mean'),
    n_participants = ('participant', 'nunique'),
    sum_item = ('n_item', 'sum')
).reset_index()

# save descriptive statistics ----

# selections
te_descriptives.to_csv('data/02_cleaned/descriptives/te_descriptives.csv', encoding='utf-8', index=False)
t_descriptives.to_csv('data/02_cleaned/descriptives/t_descriptives.csv', encoding='utf-8', index=False)
e_descriptives.to_csv('data/02_cleaned/descriptives/e_descriptives.csv', encoding='utf-8', index=False)

# rts
te_descriptives_rt.to_csv('data/02_cleaned/descriptives/te_descriptives_rt.csv', encoding='utf-8', index=False)
t_descriptives_rt.to_csv('data/02_cleaned/descriptives/t_descriptives_rt.csv', encoding='utf-8', index=False)
e_descriptives_rt.to_csv('data/02_cleaned/descriptives/e_descriptives_rt.csv', encoding='utf-8', index=False)
